R: cleanup
